added alerts

+++ portfolio-front/src/components/ItemAdmin.jsx
@@ -1,12 +1,13 @@
 import axios from "axios";
-import React, { useEffect, useState } from "react";
+import React, { useContext, useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
+import Swal from "sweetalert2";
 import ItemSearch from "./ItemSearch";
-
-import ShopNav from "./ShopNav";
+import { UserContext } from "./UserContext";
 
 export default function ItemAdmin({ items, setItems, fetchItems }) {
   const navigate = useNavigate();
+  const { handleAlert } = useContext(UserContext);
   const [itemData, setItemData] = useState({
     itemName: "",
     itemDescription: "",
@@ -34,9 +35,20 @@ export default function ItemAdmin({ items, setItems, fetchItems }) {
 
   const handleDelete = async (item) => {
     try {
-      await axios.delete(`http://localhost:3001/api/shop/items/${item._id}`);
+      const result = await Swal.fire({
+        title: "Are you sure?",
+
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes",
+      });
+      if (result.isConfirmed) {
+        await axios.delete(`http://localhost:3001/api/shop/items/${item._id}`);
+        handleAlert("success", "Successfully deleted", 1500);
+      }
     } catch (err) {
-      alert(err.message);
+      handleAlert("error", err.message, 5000);
     }
 
     fetchItems();
@@ -89,7 +101,7 @@ export default function ItemAdmin({ items, setItems, fetchItems }) {
       <div className="row pt-4">
         <div className="col-sm-4 ">
           <button
-            onClick={() => navigate("/eshop/items/addform")}
+            onClick={() => navigate("/eshop/admin/addform")}
             className="myBtn"
           >
             Add new item
@@ -104,8 +116,8 @@ export default function ItemAdmin({ items, setItems, fetchItems }) {
         <div className="col-sm-2">ITEM NAME</div>
         <div className="col-sm-2">ITEM DESCRIPTION</div>
         <div className="col-sm-2">ITEM PRICE</div>
-        <div className="col-sm-2">IMAGE PATH</div>
         <div className="col-sm-2">STOCK</div>
+        <div className="col-sm-2">IMAGE PATH</div>
         <div className="col-sm-2"></div>
       </div>
 
@@ -143,15 +155,6 @@ export default function ItemAdmin({ items, setItems, fetchItems }) {
               </div>
               <div className="col-sm-2">
                 <input
-                  name="imagePath"
-                  type="text"
-                  className="editForm inputBoxOutline"
-                  onChange={(e) => handleChange(e)}
-                  defaultValue={item.imagePath}
-                ></input>
-              </div>
-              <div className="col-sm-2">
-                <input
                   name="stock"
                   type="text"
                   className="editForm inputBoxOutline"
@@ -159,6 +162,16 @@ export default function ItemAdmin({ items, setItems, fetchItems }) {
                   defaultValue={item.stock}
                 ></input>
               </div>
+              <div className="col-sm-2">
+                <input
+                  name="imagePath"
+                  type="text"
+                  className="editForm inputBoxOutline"
+                  onChange={(e) => handleChange(e)}
+                  defaultValue={item.imagePath}
+                ></input>
+              </div>
+
               <div className="col-sm-2 alignRight">
                 <button className="myBtn3" onClick={() => handleSave(item)}>
                   Save
@@ -187,8 +200,8 @@ export default function ItemAdmin({ items, setItems, fetchItems }) {
               <div className="col-sm-2">{item.itemName}</div>
               <div className="col-sm-2">{item.itemDescription}</div>
               <div className="col-sm-2">{item.itemPrice}</div>
-              <div className="col-sm-2">{item.imagePath}</div>
               <div className="col-sm-2">{item.stock}</div>
+              <div className="col-sm-2">{item.imagePath}</div>
 
               <div className="col-sm-2 "></div>
             </div>
